Major rework of all model classes and the main BB class  * All models now implement the new version of JSONentities.  * Some progress on PR
