+++ src/main/groovy/com/eficode/atlassian/bitbucketInstanceManager/model/BitbucketUser.groovy
@@ -1,0 +1,16 @@
+package com.eficode.atlassian.bitbucketInstanceManager.model
+
+import java.nio.charset.StandardCharsets
+
+class BitbucketUser {
+
+    String name
+    String emailAddress
+
+
+    String getProfileUrl(String baseUrl) {
+
+        return baseUrl + "/users/" + URLEncoder.encode(name, StandardCharsets.UTF_8)
+    }
+}
+
